Extract main, scenario and microsoft save folder
